pyproject.toml
[project]
name = "agent13"
version = "0.1.6"
description = "Context-lean AI agent - minimal VRAM fits in a small cupboard, a postbox or even a clock"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Simon Ryan, various local models" }
]
dependencies = [
    "openai",
    "textual",
    "rich",
    "python-dotenv",
    "pyyaml",
    "pydantic",
    "pyte",
    "mcp",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "responses",
    "prompt_toolkit",
    "pexpect",
    "ruff>=0.15.12",
]

[project.scripts]
agent13 = "agent13.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent13", "ui", "tools"]
artifacts = [
    "agent13/default_skills/**",
    "agent13/default_config.toml",
    "agent13/sandbox_profiles/*.sb",
]

[tool.hatch.build.targets.sdist]
packages = ["agent13", "ui", "tools"]
artifacts = [
    "agent13/default_skills/**",
    "agent13/default_config.toml",
    "agent13/sandbox_profiles/*.sb",
]
